cosmetic: change DecalFont_DrawLineStrlen and DotLights function to use fixed point macros
redhot also edited the fixed point macro

+++ include/math.h
@@ -5,12 +5,12 @@
 // fixed point macros
 #define FRACTIONAL_BITS 12
 #define FP_ONE (1 << FRACTIONAL_BITS)
-#define FP(x) (x * FP_ONE)
+#define FP(x) ((short)(((float)x) * FP_ONE))
 
 short FP_Div(short a, short b); // see fp.c
 
 // at least one of the operands needs to be a fixed point value converted to integer form
-// e.g. FP_Mult(0x1000, 0x2000) or FP_Mult(FP(1), FP(2)) or FP_Mult(3, FP(0.75))
+// e.g. FP_Mult(0x1000, 0x2000) or FP_Mult(FP(1.0), FP(2.0)) or FP_Mult(3, FP(0.75))
 force_inline short FP_Mult(short x, short y)
 {
 	int result = x;
